[chore] updates on organization limits logic (#1755)
- cambia el tipo de dato de `integer` a `numeric` en la bbdd.
- se quitó la constraint de `used <= limit` para permitir excesos de uso
en un periodo.
- si un periodo termina con exceso de uso (`used > limit`), este se
traslada como `used` en el periodo siguiente.
- en la mutación de `sendPetition` se verifica previamente que la org
tenga créditos suficientes para enviar a todos los grupos de contactos.
(evita exceso de uso en envíos masivos)

---------

+++ client/graphql/__types.ts
@@ -2417,11 +2417,11 @@ export interface OrganizationUsageLimit {
   __typename?: "OrganizationUsageLimit";
   cycleNumber: Scalars["Int"]["output"];
   id: Scalars["GID"]["output"];
-  limit: Scalars["Int"]["output"];
+  limit: Scalars["Float"]["output"];
   period: Scalars["Duration"]["output"];
   periodEndDate?: Maybe<Scalars["DateTime"]["output"]>;
   periodStartDate: Scalars["DateTime"]["output"];
-  used: Scalars["Int"]["output"];
+  used: Scalars["Float"]["output"];
 }
 
 export type OrganizationUsageLimitName = "PETITION_SEND" | "SIGNATURIT_SHARED_APIKEY";
